+++ schema/site_stringdata.go
@@ -75,6 +75,13 @@ const SiteSchemaJSON = `{
           "enum": ["enabled", "disabled"],
           "default": "disabled"
         },
+        "configVars": {
+          "description":
+            "Enables or disables the interpolation of variables in string values in the site configuration JSON data.\n\nIf set to \"enabled\", all string values (except for JSON object property names) are interpolated, replacing ${var} or $var in the string with the value of the referenced environment variable (or the empty string if undefined).",
+          "type": "string",
+          "enum": ["enabled", "disabled"],
+          "default": "disabled"
+        },
         "multipleAuthProviders": {
           "description":
             "Enables or disables the use of multiple authentication providers and a publicly accessible web page displaying authentication options for unauthenticated users. (WARNING: Do not use this unless you know what you're doing.)",